+++ app/lib/bloc/game/game_state.dart
@@ -7,6 +7,11 @@ class GameState {
   final bool isWhiteBottom;
   final bool rotateOnMove;
   final bool mirrorTopPieces;
+  final bool showZoomOutButtonOnLeft;
+  final bool showLegalMoves;
+  final bool autoPromoteToQueen;
+  final bool allowUndo;
+  final AutoZoomOutOnMoveBehaviour autoZoomOutOnMove;
 
   const GameState({
     required this.game,
@@ -14,6 +19,11 @@ class GameState {
     required this.isWhiteBottom,
     required this.rotateOnMove,
     required this.mirrorTopPieces,
+    required this.showZoomOutButtonOnLeft,
+    required this.showLegalMoves,
+    required this.autoPromoteToQueen,
+    required this.allowUndo,
+    required this.autoZoomOutOnMove,
   });
 
   const GameState.initial({
@@ -21,6 +31,11 @@ class GameState {
     required this.isWhiteBottom,
     required this.rotateOnMove,
     required this.mirrorTopPieces,
+    required this.showZoomOutButtonOnLeft,
+    required this.showLegalMoves,
+    required this.autoPromoteToQueen,
+    required this.allowUndo,
+    required this.autoZoomOutOnMove,
   }) : enableZoomButton = false;
 
   GameState copyWith({
@@ -29,6 +44,11 @@ class GameState {
     bool? isWhiteBottom,
     bool? rotateOnMove,
     bool? mirrorTopPieces,
+    bool? showZoomOutButtonOnLeft,
+    bool? showLegalMoves,
+    bool? autoPromoteToQueen,
+    bool? allowUndo,
+    AutoZoomOutOnMoveBehaviour? autoZoomOutOnMove,
   }) {
     return GameState(
       game: game ?? this.game,
@@ -36,6 +56,12 @@ class GameState {
       isWhiteBottom: isWhiteBottom ?? this.isWhiteBottom,
       rotateOnMove: rotateOnMove ?? this.rotateOnMove,
       mirrorTopPieces: mirrorTopPieces ?? this.mirrorTopPieces,
+      showZoomOutButtonOnLeft:
+          showZoomOutButtonOnLeft ?? this.showZoomOutButtonOnLeft,
+      showLegalMoves: showLegalMoves ?? this.showLegalMoves,
+      autoPromoteToQueen: autoPromoteToQueen ?? this.autoPromoteToQueen,
+      allowUndo: allowUndo ?? this.allowUndo,
+      autoZoomOutOnMove: autoZoomOutOnMove ?? this.autoZoomOutOnMove,
     );
   }
 }