+++ app/lib/router/routes.dart
@@ -4,4 +4,5 @@ abstract class ICRoute {
   static const String home = '/home';
   static const String game = '/game';
   static const String settings = '/settings';
+  static const String settingsOtb = '/settingsOtb';
 }