+++ app/lib/widgets/ic_board.dart
@@ -5,13 +5,16 @@ import 'package:flutter_svg/svg.dart';
 import 'package:insanichess/insanichess.dart' as insanichess;
 
 import '../style/colors.dart';
+import '../style/images.dart';
 import '../util/extensions/piece.dart';
 
 class ICBoard extends StatefulWidget {
   final insanichess.Game game;
-  final void Function(insanichess.Square, insanichess.Square) onMove;
+  final void Function(insanichess.Move) onMove;
   final bool isWhiteBottom;
   final bool mirrorTopPieces;
+  final bool showLegalMoves;
+  final bool autoPromoteToQueen;
   final int scaleResetAnimationDuration;
   final Stream<void>? resetZoomStream;
   final ValueChanged<double>? onZoomChanged;
@@ -22,6 +25,8 @@ class ICBoard extends StatefulWidget {
     required this.onMove,
     required this.isWhiteBottom,
     required this.mirrorTopPieces,
+    required this.showLegalMoves,
+    required this.autoPromoteToQueen,
     this.scaleResetAnimationDuration = 0,
     this.resetZoomStream,
     this.onZoomChanged,
@@ -38,9 +43,11 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
   Animation<Matrix4>? _animationReset;
 
   StreamSubscription<void>? _resetZoomSubscription;
+  double _zoomValue = 1.0;
 
   insanichess.Square? _selectedSquare;
-  double _zoomValue = 1.0;
+  Iterable<insanichess.Square>? _legalSquaresFromSelectedSquare;
+  insanichess.Square? _squareForPromotion;
 
   @override
   void initState() {
@@ -87,37 +94,70 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
     }
   }
 
-  Widget _generateSquare(int row, int col, {required BuildContext context}) {
-    final Size screenSize = MediaQuery.of(context).size;
-    final double maxSquareSize = (screenSize.width < screenSize.height
-            ? screenSize.width
-            : screenSize.height) /
-        insanichess.Board.size;
+  void _onSquareTap(int row, int col, {required BuildContext context}) {
+    final insanichess.Square selectedSquare = insanichess.Square(row, col);
+    final insanichess.Piece? pieceOnSelectedSquare =
+        widget.game.board.at(row, col);
 
+    if (_selectedSquare == null ||
+        !_legalSquaresFromSelectedSquare!.contains(selectedSquare)) {
+      if (widget.game.playerOnTurn == insanichess.PieceColor.white &&
+              (pieceOnSelectedSquare?.isWhite ?? false) ||
+          widget.game.playerOnTurn == insanichess.PieceColor.black &&
+              (pieceOnSelectedSquare?.isBlack ?? false)) {
+        _legalSquaresFromSelectedSquare = widget.game.legalMoves
+            .where((insanichess.Move move) => move.from == _selectedSquare)
+            .map<insanichess.Square>((insanichess.Move move) => move.to);
+        setState(() => _selectedSquare = selectedSquare);
+      }
+      return;
+    }
+
+    insanichess.Piece? promoteTo;
+    final insanichess.Piece selectedPiece =
+        widget.game.board.at(_selectedSquare!.row, _selectedSquare!.col)!;
+    if ((selectedPiece is insanichess.WhitePawn &&
+            row == insanichess.Board.size - 1) ||
+        (selectedPiece is insanichess.BlackPawn && row == 0)) {
+      if (widget.autoPromoteToQueen) {
+        promoteTo = const insanichess.WhiteQueen();
+      } else {
+        setState(() => _squareForPromotion = selectedSquare);
+        return;
+      }
+    }
+
+    widget.onMove(insanichess.Move(
+      _selectedSquare!,
+      selectedSquare,
+      promoteTo,
+    ));
+    _legalSquaresFromSelectedSquare = null;
+    setState(() => _selectedSquare = null);
+  }
+
+  Widget _generateSquare(
+    int row,
+    int col, {
+    required double maxSquareSize,
+    required BuildContext context,
+  }) {
     final bool shouldMirror = !widget.mirrorTopPieces
         ? false
         : (widget.isWhiteBottom
             ? (widget.game.board.at(row, col)?.isBlack ?? false)
             : (widget.game.board.at(row, col)?.isWhite ?? false));
 
+    final bool hasLegalMoveIndicator = !widget.showLegalMoves
+        ? false
+        : _legalSquaresFromSelectedSquare
+                ?.contains(insanichess.Square(row, col)) ??
+            false;
+
     return GestureDetector(
       onTap: widget.game.isGameOver || widget.game.canGoForward
           ? null
-          : () {
-              if (_selectedSquare == null) {
-                if (widget.game.playerOnTurn == insanichess.PieceColor.white &&
-                        (widget.game.board.at(row, col)?.isWhite ?? false) ||
-                    widget.game.playerOnTurn == insanichess.PieceColor.black &&
-                        (widget.game.board.at(row, col)?.isBlack ?? false)) {
-                  setState(
-                      () => _selectedSquare = insanichess.Square(row, col));
-                }
-                return;
-              }
-
-              widget.onMove(_selectedSquare!, insanichess.Square(row, col));
-              setState(() => _selectedSquare = null);
-            },
+          : () => _onSquareTap(row, col, context: context),
       child: Container(
         height: maxSquareSize, // might change
         width: maxSquareSize, // might change
@@ -135,13 +175,30 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
                   : ICColor.chessboardWhite,
         ),
         child: widget.game.board.at(row, col) == null
-            ? const SizedBox.expand()
-            : SvgPicture.asset(
-                widget.game.board
-                    .at(row, col)!
-                    .getImagePath(mirrored: shouldMirror),
-                width: maxSquareSize,
-                height: maxSquareSize,
+            ? (hasLegalMoveIndicator
+                ? Icon(
+                    CupertinoIcons.circle_fill,
+                    size: maxSquareSize / 3,
+                    color: const Color(0xdd888888),
+                  )
+                : null)
+            : Stack(
+                alignment: Alignment.center,
+                children: <Widget>[
+                  SvgPicture.asset(
+                    widget.game.board
+                        .at(row, col)!
+                        .getImagePath(mirrored: shouldMirror),
+                    width: maxSquareSize,
+                    height: maxSquareSize,
+                  ),
+                  if (hasLegalMoveIndicator)
+                    Icon(
+                      CupertinoIcons.circle_fill,
+                      size: maxSquareSize / 3,
+                      color: const Color(0xdd888888),
+                    ),
+                ],
               ),
       ),
     );
@@ -149,13 +206,24 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
+    final Size screenSize = MediaQuery.of(context).size;
+    final double maxSquareSize = (screenSize.width < screenSize.height
+            ? screenSize.width
+            : screenSize.height) /
+        insanichess.Board.size;
+
     final List<Row> rows = <Row>[];
 
     if (widget.isWhiteBottom) {
       for (int row = insanichess.Board.size - 1; row >= 0; row--) {
         final List<Widget> children = <Widget>[];
         for (int col = 0; col < insanichess.Board.size; col++) {
-          children.add(_generateSquare(row, col, context: context));
+          children.add(_generateSquare(
+            row,
+            col,
+            context: context,
+            maxSquareSize: maxSquareSize,
+          ));
         }
         rows.add(Row(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -166,13 +234,90 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
       for (int row = 0; row < insanichess.Board.size; row++) {
         final List<Widget> children = <Widget>[];
         for (int col = insanichess.Board.size - 1; col >= 0; col--) {
-          children.add(_generateSquare(row, col, context: context));
+          children.add(_generateSquare(
+            row,
+            col,
+            context: context,
+            maxSquareSize: maxSquareSize,
+          ));
         }
         rows.add(Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: children,
         ));
       }
+    }
+
+    Widget child = Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: rows,
+    );
+    if (_squareForPromotion != null) {
+      final bool isWhitePromoting =
+          _squareForPromotion!.row == insanichess.Board.size - 1;
+      final double? top;
+      final double? bottom;
+      final double? left;
+
+      if (isWhitePromoting) {
+        if (widget.isWhiteBottom) {
+          top = 0.0;
+          bottom = null;
+          left = _squareForPromotion!.col * maxSquareSize;
+        } else {
+          top = null;
+          bottom = 0.0;
+          left = (insanichess.Board.size - 2 - _squareForPromotion!.col) *
+              maxSquareSize;
+        }
+      } else {
+        if (widget.isWhiteBottom) {
+          top = null;
+          bottom = 0.0;
+          left = (insanichess.Board.size - 2 - _squareForPromotion!.col) *
+              maxSquareSize;
+        } else {
+          top = 0.0;
+          bottom = null;
+          left = _squareForPromotion!.col * maxSquareSize;
+        }
+      }
+
+      final bool queenTop = (isWhitePromoting && widget.isWhiteBottom) ||
+          (!isWhitePromoting && !widget.isWhiteBottom);
+
+      child = Stack(
+        children: <Widget>[
+          child,
+          Positioned(
+            top: top,
+            bottom: bottom,
+            left: left,
+            child: _PromotionSelectionOverlay(
+              isWhitePromoting: isWhitePromoting,
+              queenTop: queenTop,
+              // Queen is bottom when top player is promoting, therefore pieces
+              // are only mirrored when queen is not top and top pieces should
+              // be mirrored.
+              mirrorPieces: !queenTop && widget.mirrorTopPieces,
+              onPromotionSelected: (insanichess.Piece promoteTo) {
+                print('promotion selected');
+                widget.onMove(insanichess.Move(
+                  _selectedSquare!,
+                  _squareForPromotion!,
+                  promoteTo,
+                ));
+                setState(() {
+                  _selectedSquare = null;
+                  _squareForPromotion = null;
+                  _legalSquaresFromSelectedSquare = null;
+                });
+              },
+            ),
+          ),
+        ],
+      );
     }
 
     return InteractiveViewer(
@@ -197,10 +342,170 @@ class _ICBoardState extends State<ICBoard> with TickerProviderStateMixin {
           _zoomValue = newZoomValue;
         }
       },
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: rows,
+      child: child,
+    );
+  }
+}
+
+class _PromotionSelectionOverlay extends StatelessWidget {
+  final bool isWhitePromoting;
+  final bool queenTop;
+  final bool mirrorPieces;
+  final ValueChanged<insanichess.Piece> onPromotionSelected;
+
+  const _PromotionSelectionOverlay({
+    Key? key,
+    required this.isWhitePromoting,
+    required this.queenTop,
+    required this.mirrorPieces,
+    required this.onPromotionSelected,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final double width =
+        MediaQuery.of(context).size.width / insanichess.Board.size;
+
+    List<Widget> children;
+    if (isWhitePromoting) {
+      children = mirrorPieces
+          ? <Widget>[
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteQueenR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteQueen()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteRookR,
+                size: width,
+                onTap: () => onPromotionSelected(const insanichess.WhiteRook()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteKnightR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteKnight()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteBishopR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteBishop()),
+              ),
+            ]
+          : <Widget>[
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteQueen,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteQueen()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteRook,
+                size: width,
+                onTap: () => onPromotionSelected(const insanichess.WhiteRook()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteKnight,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteKnight()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceWhiteBishop,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.WhiteBishop()),
+              ),
+            ];
+    } else {
+      children = mirrorPieces
+          ? <Widget>[
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackQueenR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackQueen()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackRookR,
+                size: width,
+                onTap: () => onPromotionSelected(const insanichess.BlackRook()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackKnightR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackKnight()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackBishopR,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackBishop()),
+              ),
+            ]
+          : <Widget>[
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackQueen,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackQueen()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackRook,
+                size: width,
+                onTap: () => onPromotionSelected(const insanichess.BlackRook()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackKnight,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackKnight()),
+              ),
+              _PromotionPieceButton(
+                imagePath: ICImage.pieceBlackBishop,
+                size: width,
+                onTap: () =>
+                    onPromotionSelected(const insanichess.BlackBishop()),
+              ),
+            ];
+    }
+    if (!queenTop) {
+      children = children.reversed.toList(growable: false);
+    }
+
+    return Container(
+      width: width,
+      height: 4 * width,
+      color: const Color(0xffffffff),
+      child: Column(children: children),
+    );
+  }
+}
+
+class _PromotionPieceButton extends StatelessWidget {
+  final String imagePath;
+  final double size;
+  final VoidCallback onTap;
+
+  const _PromotionPieceButton({
+    Key? key,
+    required this.imagePath,
+    required this.size,
+    required this.onTap,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      behavior: HitTestBehavior.opaque,
+      onTap: onTap,
+      child: SvgPicture.asset(
+        imagePath,
+        width: size,
+        height: size,
       ),
     );
   }