+++ packages/insanichess_sdk/lib/src/models/settings/game/game.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/foundation.dart';
 
+import '../../../util/enum/auto_zoom_out_on_move_behaviour.dart';
+
 /// Model for game settings.
 class InsanichessGameSettings {
   /// Is undoing / takeback allowed?
@@ -8,23 +10,44 @@ class InsanichessGameSettings {
   /// Should pawns automatically be promoted to queen?
   final bool alwaysPromoteToQueen;
 
+  /// Should the board be zoomed out automatically on move, and if yes, how.
+  final AutoZoomOutOnMoveBehaviour autoZoomOutOnMove;
+
   /// Constructs new `InsanichessGameSettings` object with [allowUndo] and
   /// [alwaysPromoteToQueen].
   const InsanichessGameSettings({
     required this.allowUndo,
     required this.alwaysPromoteToQueen,
+    required this.autoZoomOutOnMove,
   });
 
   /// Constructs new `InsanichessGameSettings` object with default values.
   const InsanichessGameSettings.defaults()
       : allowUndo = true,
-        alwaysPromoteToQueen = false;
+        alwaysPromoteToQueen = false,
+        autoZoomOutOnMove = AutoZoomOutOnMoveBehaviour.always;
 
   /// Constructs new `InsanichessGameSettings` object from [json].
   InsanichessGameSettings.fromJson(Map<String, dynamic> json)
       : allowUndo = json[InsanichessGameSettingsJsonKey.allowUndo],
         alwaysPromoteToQueen =
-            json[InsanichessGameSettingsJsonKey.alwaysPromoteToQueen];
+            json[InsanichessGameSettingsJsonKey.alwaysPromoteToQueen],
+        autoZoomOutOnMove = autoZoomOutOnMoveBehaviourFromJson(
+            json[InsanichessGameSettingsJsonKey.autoZoomOutOnMove]);
+
+  /// Returns a new `InsanichessGameSettings` object by overriding existing
+  /// field values with those given in arguments.
+  InsanichessGameSettings copyWith({
+    bool? allowUndo,
+    bool? alwaysPromoteToQueen,
+    AutoZoomOutOnMoveBehaviour? autoZoomOutOnMove,
+  }) {
+    return InsanichessGameSettings(
+      allowUndo: allowUndo ?? this.allowUndo,
+      alwaysPromoteToQueen: alwaysPromoteToQueen ?? this.alwaysPromoteToQueen,
+      autoZoomOutOnMove: autoZoomOutOnMove ?? this.autoZoomOutOnMove,
+    );
+  }
 
   /// Converts this object to json representation.
   @mustCallSuper
@@ -32,6 +55,8 @@ class InsanichessGameSettings {
     return <String, dynamic>{
       InsanichessGameSettingsJsonKey.allowUndo: allowUndo,
       InsanichessGameSettingsJsonKey.alwaysPromoteToQueen: alwaysPromoteToQueen,
+      InsanichessGameSettingsJsonKey.autoZoomOutOnMove:
+          autoZoomOutOnMove.toJson(),
     };
   }
 }
@@ -43,4 +68,7 @@ abstract class InsanichessGameSettingsJsonKey {
 
   /// Key for `InsanichessGameSettings.alwaysPromoteToQueen`.
   static const String alwaysPromoteToQueen = 'always_promote_to_queen';
+
+  /// Key for `InsanichessSettings.autoZoomOutOnMove`.
+  static const String autoZoomOutOnMove = 'auto_zoom_out_on_move';
 }