+++ packages/insanichess_sdk/lib/src/models/settings/game/otb.dart
@@ -1,3 +1,4 @@
+import '../../../../insanichess_sdk.dart';
 import 'game.dart';
 
 /// Model for OTB settings.
@@ -16,9 +17,11 @@ class InsanichessOtbSettings extends InsanichessGameSettings {
     required this.mirrorTopPieces,
     required bool allowUndo,
     required bool alwaysPromoteToQueen,
+    required AutoZoomOutOnMoveBehaviour autoZoomOutOnMove,
   }) : super(
           allowUndo: allowUndo,
           alwaysPromoteToQueen: alwaysPromoteToQueen,
+          autoZoomOutOnMove: autoZoomOutOnMove,
         );
 
   /// Creates new `InsanichessOtbSettings` object with default values.
@@ -32,6 +35,25 @@ class InsanichessOtbSettings extends InsanichessGameSettings {
       : rotateChessboard = json[InsanichessOtbSettingsJsonKey.rotateChessboard],
         mirrorTopPieces = json[InsanichessOtbSettingsJsonKey.mirrorTopPieces],
         super.fromJson(json);
+
+  /// Returns a new `InsanichessOtbSettings` object by overriding existing field
+  /// values with those given in arguments.
+  @override
+  InsanichessOtbSettings copyWith({
+    bool? rotateChessboard,
+    bool? mirrorTopPieces,
+    bool? allowUndo,
+    bool? alwaysPromoteToQueen,
+    AutoZoomOutOnMoveBehaviour? autoZoomOutOnMove,
+  }) {
+    return InsanichessOtbSettings(
+      rotateChessboard: rotateChessboard ?? this.rotateChessboard,
+      mirrorTopPieces: mirrorTopPieces ?? this.mirrorTopPieces,
+      allowUndo: allowUndo ?? this.allowUndo,
+      alwaysPromoteToQueen: alwaysPromoteToQueen ?? this.alwaysPromoteToQueen,
+      autoZoomOutOnMove: autoZoomOutOnMove ?? this.autoZoomOutOnMove,
+    );
+  }
 
   /// Returns json representation of this object.
   @override