+++ packages/insanichess_sdk/lib/src/models/settings/settings.dart
@@ -1,4 +1,3 @@
-import '../../util/enum/auto_zoom_out_on_move_behaviour.dart';
 import 'game/otb.dart';
 
 /// Model for all settings for the app.
@@ -12,29 +11,18 @@ class InsanichessSettings {
   /// When a piece is selected, should we show legal moves with that piece.
   final bool showLegalMoves;
 
-  /// If a piece is selected and only one legal move can be played with it,
-  /// should that move be automatically performed?
-  final bool autoMoveIfOneLegalMove;
-
-  /// Should the board be zoomed out automatically on move, and if yes, how.
-  final AutoZoomOutOnMoveBehaviour autoZoomOutOnMove;
-
   /// Creates new `InsanichessSettings` object.
   const InsanichessSettings({
     required this.otb,
     required this.showZoomOutButtonOnLeft,
     required this.showLegalMoves,
-    required this.autoMoveIfOneLegalMove,
-    required this.autoZoomOutOnMove,
   });
 
   /// Creates new `InsanichessSettings` object with default values.
   const InsanichessSettings.defaults()
       : otb = const InsanichessOtbSettings.defaults(),
         showZoomOutButtonOnLeft = true,
-        showLegalMoves = true,
-        autoMoveIfOneLegalMove = true,
-        autoZoomOutOnMove = AutoZoomOutOnMoveBehaviour.onMyMove;
+        showLegalMoves = true;
 
   /// Creates new `InsanichessSettings` from [json] representation.
   InsanichessSettings.fromJson(Map<String, dynamic> json)
@@ -42,11 +30,22 @@ class InsanichessSettings {
             json[InsanichessSettingsJsonKey.otb]),
         showZoomOutButtonOnLeft =
             json[InsanichessSettingsJsonKey.showZoomOutButtonOnLeft],
-        showLegalMoves = json[InsanichessSettingsJsonKey.showLegalMoves],
-        autoMoveIfOneLegalMove =
-            json[InsanichessSettingsJsonKey.autoMoveIfOneLegalMove],
-        autoZoomOutOnMove = autoZoomOutOnMoveBehaviourFromJson(
-            json[InsanichessSettingsJsonKey.autoZoomOutOnMove]);
+        showLegalMoves = json[InsanichessSettingsJsonKey.showLegalMoves];
+
+  /// Returns a new `InsanichessSettings` object by overriding existing field
+  /// values with those given in arguments.
+  InsanichessSettings copyWith({
+    InsanichessOtbSettings? otb,
+    bool? showZoomOutButtonOnLeft,
+    bool? showLegalMoves,
+  }) {
+    return InsanichessSettings(
+      otb: otb ?? this.otb,
+      showZoomOutButtonOnLeft:
+          showZoomOutButtonOnLeft ?? this.showZoomOutButtonOnLeft,
+      showLegalMoves: showLegalMoves ?? this.showLegalMoves,
+    );
+  }
 
   /// Converts this object to json representation.
   Map<String, dynamic> toJson() {
@@ -55,8 +54,6 @@ class InsanichessSettings {
       InsanichessSettingsJsonKey.showZoomOutButtonOnLeft:
           showZoomOutButtonOnLeft,
       InsanichessSettingsJsonKey.showLegalMoves: showLegalMoves,
-      InsanichessSettingsJsonKey.autoMoveIfOneLegalMove: autoMoveIfOneLegalMove,
-      InsanichessSettingsJsonKey.autoZoomOutOnMove: autoZoomOutOnMove.toJson(),
     };
   }
 }
@@ -71,10 +68,4 @@ abstract class InsanichessSettingsJsonKey {
 
   /// Key for `InsanichessSettings.showLegalMoves`.
   static const String showLegalMoves = 'show_legal_moves';
-
-  /// Key for `InsanichessSettings.autoMoveIfOneLegalMove`.
-  static const String autoMoveIfOneLegalMove = 'auto_move_if_one_legal_move';
-
-  /// Key for `InsanichessSettings.autoZoomOutOnMove`.
-  static const String autoZoomOutOnMove = 'auto_zoom_out_on_move';
 }